+++ CHANGELOG.md
@@ -6,10 +6,12 @@ occur without warning between the alpha-releases. The official release will have
 ## 0.1.0-alpha13
 
 * Major restructuring: (`scripts/migrate-to-new-package-structure.sh` can help rewrite your code to new structure)
-  * `control-flow`, `git`, `manualtrigger` and `shell` are now in `steps`-namespace
+  * `control-flow`, `git`, `manualtrigger` and `shell` are now in `steps`-package
   * `lambdacd.presentation` is now `lambdacd.presentation.pipeline-structure`
   * some functions more concerned with presentation than actual management of the pipeline state are moved from `lambdacd.pipeline-state` to `lambdacd.presentation.pipeline-state`
-
+  * internals are now in the `internal` package
+  * ui related functionality is now in the `ui` package
+  
 ## 0.1.0-alpha12
 
 ## 0.1.0-alpha11