+++ CHANGELOG.md
@@ -5,7 +5,7 @@ occur without warning between the alpha-releases. The official release will have
 
 ## 0.1.0-alpha13
 
-* Major restructuring: (`scripts/migrate-to-new-package-structure.sh` can help rewrite your code to new structure)
+* Major restructuring: (`scripts/migrate-to-new-package-structure.sh` can help rewrite your code to work with the new structure)
   * `control-flow`, `git`, `manualtrigger` and `shell` are now in `steps`-package
   * `lambdacd.presentation` is now `lambdacd.presentation.pipeline-structure`
   * some functions more concerned with presentation than actual management of the pipeline state are moved from `lambdacd.pipeline-state` to `lambdacd.presentation.pipeline-state`