+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 This changelog indicates breaking changes to the API. As this is still an experimental library, breaking changes may
 occur without warning between the alpha-releases. The official release will have a defined and more stable API. 
 
+## 0.1.0-alpha14 [wip]
+
+* LambdaCD now requires Clojure 1.6.0
+
 ## 0.1.0-alpha13
 
 * Major restructuring: (`scripts/migrate-to-new-package-structure.sh` can help rewrite your code to work with the new structure)